--- a/tests/deployment/test_deployment_readiness.py
+++ b/tests/deployment/test_deployment_readiness.py
@@ -1,0 +1,1 @@
+# Deployment readiness tests