--- a/tests/scripts/test_health_check.py
+++ b/tests/scripts/test_health_check.py
@@ -1,0 +1,1 @@
+# Health check tests