… .github/workflows/ci.yml
… .github/workflows/deploy.yml
… .github/workflows/test.yml
… docker/docker-compose.yml
… project-documents/commit-messages/commit_1_2_1.md
… project-documents/enhancements/enhancements_1_2.md
… project-documents/punchlist.md
… scripts/health-check.py
… tests/integration/test_full_structure.py
… tests/structure/test_conftest_fixture_validation.py
… tests/structure/test_cross_directory_imports.py
… tests/structure/test_file_encoding_format.py
… tests/structure/test_project_performance.py
… tests/structure/test_scalability.py